backend/pyproject.toml
[tool.poetry]
name = "commwell"
version = "0.1.0"
description = ""
authors = ["Igor Klepuszewski <igor.klepuszewski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pytest-django = "^4.9.0"
psycopg2 = "^2.9.9"
pillow = "^10.4.0"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
